requirements.txt
streamlit
langchain
langchain-groq
tavily-python
pdfplumber
python-dotenv
tiktoken
